Change item pack size + quantity inserts back to whole number
Source: https://github.com/msupply-foundation/open-msupply/commit/fe7b3a5c0de2c086653a86ac017f80c6808d1774

+++ server/graphql/invoice_line/src/mutations/outbound_shipment_line/unallocated_line/insert.rs
@@ -19,7 +19,7 @@ pub struct InsertInput {
     pub id: String,
     pub invoice_id: String,
     pub item_id: String,
-    pub quantity: f64,
+    pub quantity: u32,
 }
 
 #[derive(Interface)]
