+++ server/service/src/invoice_line/outbound_shipment_unallocated_line/update.rs
@@ -10,7 +10,7 @@ use repository::{
 #[derive(Clone, Debug, PartialEq, Default)]
 pub struct UpdateOutboundShipmentUnallocatedLine {
     pub id: String,
-    pub quantity: f64,
+    pub quantity: u32,
 }
 
 #[derive(Clone, Debug, PartialEq)]
@@ -70,7 +70,7 @@ fn generate(
     }: UpdateOutboundShipmentUnallocatedLine,
     mut line: InvoiceLineRow,
 ) -> Result<InvoiceLineRow, UpdateOutboundShipmentUnallocatedLineError> {
-    line.number_of_packs = quantity;
+    line.number_of_packs = quantity as f64;
 
     Ok(line)
 }
@@ -123,7 +123,7 @@ mod test_update {
                 &context,
                 UpdateOutboundShipmentUnallocatedLine {
                     id: "invalid".to_owned(),
-                    quantity: 0.0
+                    quantity: 0
                 },
             ),
             Err(ServiceError::LineDoesNotExist)
@@ -135,7 +135,7 @@ mod test_update {
                 &context,
                 UpdateOutboundShipmentUnallocatedLine {
                     id: mock_outbound_shipment_a_invoice_lines()[0].id.clone(),
-                    quantity: 0.0
+                    quantity: 0
                 },
             ),
             Err(ServiceError::LineIsNotUnallocatedLine)
@@ -147,7 +147,7 @@ mod test_update {
                 &context,
                 UpdateOutboundShipmentUnallocatedLine {
                     id: mock_unallocated_line().id,
-                    quantity: 0.0
+                    quantity: 0
                 },
             ),
             Err(ServiceError::NotThisStoreInvoice)
@@ -173,7 +173,7 @@ mod test_update {
                 &context,
                 UpdateOutboundShipmentUnallocatedLine {
                     id: line_to_update.id.clone(),
-                    quantity: 20.0,
+                    quantity: 20,
                 },
             )
             .unwrap();
